trobz_dashboard: fix few bugs: numeric values use float, link to metric list only if model is defined

+++ trobz_dashboard/static/src/js/views/widgets/display/numeric.js
@@ -22,7 +22,7 @@ openerp.trobz.module('trobz_dashboard',function(dashboard, _, Backbone, base){
         serializeData: function(){
             
             var process_value = function(value, options){
-                return numeral(parseInt(value)).format(options.format);
+                return numeral(parseFloat(value)).format(options.format);
             };
             
             var process_thresholders = function(value, options){
